# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Platform Wallet Point Gamifikasi Kampus - Backend API",
        "version": "1.0",
        "title": "Wallet Point API",
        "contact": {
            "name": "API Support",
            "email": "support@campus.edu"
        }
    },
    "host": "localhost:8102",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "Enter your bearer token in the format: Bearer <token>"
        }
    },
    "paths": {
...
        "/mahasiswa/missions/submit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Mahasiswa - Missions"
                ],
                "summary": "Submit Mission",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "mission_id": {
                                    "type": "integer"
                                },
                                "content": {
                                    "type": "string"
                                },
                                "file_url": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submitted"
                    }
                }
            }
        },
...
        "/dosen/missions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Dosen - Missions"
                ],
                "summary": "Create Mission",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "quiz",
                                        "task",
                                        "assignment"
                                    ]
                                },
                                "points_reward": {
                                    "type": "number"
                                },
                                "deadline": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Mission created"
                    }
                }
            },
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Dosen - Missions"
                ],
                "summary": "Get My Missions",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}